Update README.md
I made a google bar. I did with CSS and HTML. It is not a dynamic page because I didn't use any JS code or backend code. It is just front.
I hope you enjoy it and I'll be happy if you comment your oppinion and help me to do this projects better.